+++ test/test_conditional.py
@@ -106,7 +106,7 @@ print('====')
 
 program='''
 wait_sync 100
-latch_en 1,4
+set_latch_en 1,4
 #Q1Sim:sim_trigger 1, 1
 play 0,0,84 # subtract 16ns: total of else_wait of if/elif conditionals (final else not included)